Split Rust query helper types
